row: move Metrics struct into rowinfra
This allows us to break dependency of `execinfra` on `row` which speeds
up the build a bit.

Release note: None
