+++ pkg/sql/execinfra/BUILD.bazel
@@ -47,9 +47,9 @@ go_library(
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/execinfrapb",
-        "//pkg/sql/row",
         "//pkg/sql/rowenc",
         "//pkg/sql/rowenc/valueside",
+        "//pkg/sql/rowinfra",
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
@@ -123,6 +123,7 @@ disallowed_imports_test(
         "//pkg/sql/colexec",
         "//pkg/sql/colflow",
         "//pkg/sql/flowinfra",
+        "//pkg/sql/row",
         "//pkg/sql/rowexec",
         "//pkg/sql/rowflow",
     ],
