+++ pkg/sql/execinfra/server_config.go
@@ -31,7 +31,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/rpc/nodedialer"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
-	"github.com/cockroachdb/cockroach/pkg/sql/row"
+	"github.com/cockroachdb/cockroach/pkg/sql/rowinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlliveness"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlutil"
@@ -120,8 +120,8 @@ type ServerConfig struct {
 	ParentDiskMonitor *mon.BytesMonitor
 
 	Metrics            *DistSQLMetrics
-	RowMetrics         *row.Metrics
-	InternalRowMetrics *row.Metrics
+	RowMetrics         *rowinfra.Metrics
+	InternalRowMetrics *rowinfra.Metrics
 
 	// SQLLivenessReader provides access to reading the liveness of sessions.
 	SQLLivenessReader sqlliveness.Reader
@@ -316,9 +316,9 @@ func GetWorkMemLimit(flowCtx *FlowCtx) int64 {
 	return flowCtx.EvalCtx.SessionData().WorkMemLimit
 }
 
-// GetRowMetrics returns the proper RowMetrics for either internal or user
+// GetRowMetrics returns the proper rowinfra.Metrics for either internal or user
 // queries.
-func (flowCtx *FlowCtx) GetRowMetrics() *row.Metrics {
+func (flowCtx *FlowCtx) GetRowMetrics() *rowinfra.Metrics {
 	if flowCtx.EvalCtx.SessionData().Internal {
 		return flowCtx.Cfg.InternalRowMetrics
 	}
