+++ pkg/sql/row/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "kv_batch_streamer.go",
         "kv_fetcher.go",
         "locking.go",
-        "metrics.go",
         "partial_index.go",
         "row_converter.go",
         "truncate.go",
@@ -66,7 +65,6 @@ go_library(
         "//pkg/util/hlc",
         "//pkg/util/log",
         "//pkg/util/log/eventpb",
-        "//pkg/util/metric",
         "//pkg/util/mon",
         "//pkg/util/protoutil",
         "//pkg/util/timeutil",
