+++ pkg/sql/rowinfra/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "rowinfra",
-    srcs = ["base.go"],
+    srcs = [
+        "base.go",
+        "metrics.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/rowinfra",
     visibility = ["//visibility:public"],
+    deps = ["//pkg/util/metric"],
 )
