+++ pkg/sql/rowinfra/metrics.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package row
+package rowinfra
 
 import "github.com/cockroachdb/cockroach/pkg/util/metric"
